+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ This version adds tap behavior to the margin next to character inputs.
 
 ### 🗑 Deprecations
 
+* `iPhoneKeyboardLayoutProvider` has new `lowerLeadingActions` and `lowerTrailingActions` that take the row as parameter to resolve the character margin.
 * `SystemKeyboardLayoutProvider` has been redesigned to let all functions that take multiple parameters with an initial context, place the context parameter last. This gives a cleaner and more harmonized public api.   
 * `SystemKeyboardLayoutProvider` `inputs(for:)` has been renamed to `inputRows(for:)`.